rename routers & template
