+++ src/router/RMEHashRouter.js
@@ -1,5 +1,5 @@
-import RMERouterUtils from './routerUtils';
-import RMERouterContext from './routerContext';
+import RMERouterUtils from './RouterUtils';
+import RMERouterContext from './RouterContext';
 import Util from '../util';
 
 /**
