+++ src/router/functions.js
@@ -1,8 +1,8 @@
 import Component from '../component';
-import RMEHashRouter from './hashRouter';
-import RMEUrlRouter from './urlRouter';
-import RMEOnLoadUrlRouter from './onLoadUrlRouter';
-import RMERouterContext from './routerContext';
+import RMEHashRouter from './RMEHashRouter';
+import RMEUrlRouter from './RMEUrlRouter';
+import RMEOnLoadUrlRouter from './RMEOnLoadUrlRouter';
+import RMERouterContext from './RouterContext';
 import Util from '../util';
